Refactor the reserved instances cache for GCP
